使用unsigned代替long减少内存开销

+++ 278.第一个错误的版本.cpp
@@ -11,10 +11,10 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        // n可能是最大的int，用long来防止溢出
-        long low = 0, high = (long)n + 1;
+        // n可能是最大的int，用unsigned来防止溢出
+        unsigned low = 0, high = (unsigned)n + 1;
         while (high - low != 1) {
-            long mid = (low + high) >> 1;
+            unsigned mid = (low + high) >> 1;
             // 规定high所在的版本必定是bad
             if (isBadVersion((int)mid)) {
                 high = mid;
